Manual door mode from door lines

+++ src/Edgar/Legacy/Core/LayoutGenerators/PlatformersGenerator/PlatformersGenerator.cs
@@ -57,7 +57,7 @@ namespace Edgar.Legacy.Core.LayoutGenerators.PlatformersGenerator
         }
 
         // TODO: remove
-        public double TimeTotal => generator.TimeTotal;
+        public long TimeTotal => generator.TimeTotal;
 
         public int IterationsCount => generator.IterationsCount;
